# config/openapi/openapi.json
{"openapi":"3.0.0","paths":{"/items":{"get":{"operationId":"ItemsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Item"}}}}}},"summary":"List all items","tags":["items"],"x-infra-protected":false},"post":{"operationId":"ItemsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateItemDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}}}},"security":[{"bearer":[]}],"summary":"Create item","tags":["items"],"x-infra-protected":true,"x-infra-scopes":["write:items"]}},"/items/{id}":{"get":{"operationId":"ItemsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Item"}}}},"404":{"description":"Not found"}},"summary":"Get item by ID","tags":["items"],"x-infra-protected":false},"delete":{"operationId":"ItemsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"204":{"description":""},"404":{"description":"Not found"}},"security":[{"bearer":[]}],"summary":"Delete item","tags":["items"],"x-infra-protected":true,"x-infra-scopes":["write:items"]}}},"info":{"title":"Service A","description":"Demo API service","version":"v1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"Item":{"type":"object","properties":{"id":{"type":"number","example":1},"name":{"type":"string","example":"Widget"},"description":{"type":"string","example":"A useful widget"}},"required":["id","name"]},"CreateItemDto":{"type":"object","properties":{"name":{"type":"string","example":"Widget"},"description":{"type":"string","example":"A useful widget"}},"required":["name"]}}}}